final commit

+++ src/main/java/com/cs2i/libraryapi/dto/LoginResponse.java
@@ -12,5 +12,5 @@ public class LoginResponse {
     private String role;
     private String nom;
     private String prenom;
-    private float caution;
+    private double caution;
 }
